refactor(Home) : divide Home as components
