+++ src/components/Home/style/IndexList.scss
@@ -1,0 +1,8 @@
+.trending-news-list {
+	width: 20%;
+	height: 100%;
+
+	@media screen and (max-width: 820px) {
+		margin-left: 0px;
+	}
+}
