+++ src/components/Stock/StockLineChart.jsx
@@ -3,7 +3,7 @@ import images from 'assets/images';
 import ReactApexChart from 'react-apexcharts';
 import './css/StockDetailChart.scss';
 
-const StockLineChart = ({ chartData, setIsLineChart, canCandle }) => {
+const StockLineChart = ({ chartData }) => {
 	const [updateDate, setUpdateDate] = useState(0);
 	const [lineInfo, setLineInfo] = useState([]);
