+++ src/components/Stock/css/StockTileList.scss
@@ -1,5 +1,5 @@
 .tile-list {
-	width: 35%;
+	width: 100%;
 	height: 100%;
 	font-size: 0.7vw;
 	display: flex;
