+++ src/pages/Home/style/index.scss
@@ -13,44 +13,6 @@
 		padding: 24px 24px;
 	}
 
-	.home-chart-view {
-		width: 70%;
-		height: 100%;
-		display: flex;
-		flex-direction: row;
-
-		@media screen and (max-width: 820px) {
-			width: 100%;
-			flex-direction: column-reverse;
-		}
-
-		.main-chart {
-			width: 65%;
-			height: 100%;
-			display: flex;
-			flex-direction: column;
-			margin-left: 10px;
-
-			@media screen and (max-width: 820px) {
-				margin-left: 0px;
-			}
-
-			.main-title {
-				background-color: white;
-				margin-bottom: 10px;
-
-				@media screen and (max-width: 820px) {
-					font-size: 2em;
-				}
-			}
-
-			@media screen and (max-width: 820px) {
-				width: 100%;
-				height: 400px;
-			}
-		}
-	}
-
 	.line {
 		width: 1px;
 		background-color: #0c2a6b;
@@ -59,112 +21,6 @@
 
 		@media screen and (max-width: 820px) {
 			display: none;
-		}
-	}
-
-	.right-area {
-		width: 30%;
-		display: flex;
-		flex-direction: column;
-		background-color: white;
-		padding: 20px;
-
-		@media screen and (max-width: 820px) {
-			margin-top: 10px;
-			width: 100%;
-		}
-
-		.side-title {
-			button {
-				margin-right: 10px;
-				font-size: 1.2em;
-				padding: 5px 10px;
-				border-radius: 5px;
-				font-weight: bold;
-
-				@media screen and (max-width: 820px) {
-					font-size: 3em;
-				}
-			}
-
-			.unselected {
-				&:hover {
-					background-color: rgb(233, 233, 233);
-				}
-			}
-
-			.selected {
-				background-color: #0084ff;
-				color: white;
-			}
-		}
-
-		.stock-list {
-			height: 100%;
-			overflow-y: hidden;
-			font-size: 0.65em;
-			margin-top: 10px;
-			padding-right: 10x;
-
-			@media screen and (max-width: 820px) {
-				width: 100%;
-				overflow-y: auto;
-			}
-
-			a {
-				display: flex;
-				align-items: center;
-				height: calc(100% / 5);
-				border-bottom: 1px solid #cecece;
-
-				&:first-child {
-					border-top: 1px solid #cecece;
-				}
-
-				&:hover {
-					background-color: rgb(227, 244, 253);
-				}
-			}
-
-			.trending-item {
-				width: 100%;
-				display: flex;
-				text-align: left;
-
-				@media screen and (max-width: 820px) {
-					font-size: 2.2em;
-				}
-			}
-
-			.news-item {
-				width: 100%;
-				display: flex;
-				justify-content: space-between;
-				align-items: center;
-				padding: 10px;
-
-				div {
-					width: 100%;
-					display: flex;
-					flex-direction: column;
-					text-align: left;
-					overflow: hidden;
-
-					.news-title {
-						font-size: 1.7em;
-						font-weight: bold;
-						padding-right: 20px;
-					}
-
-					.news-date {
-						text-align: left;
-					}
-				}
-
-				img {
-					width: 25%;
-				}
-			}
 		}
 	}
 }
